fix: markdownlint issue with proper-names
There is no near-term way to add support for Hugo-style links to markdownlint to avoid the problem with MD044. There is a hacky workaround but it is too hacky and it is better to just ignore the lines, which this PR does.

See https://github.com/DavidAnson/markdownlint/issues/1081#issuecomment-1870229958